server/requirements.txt
fastapi
uvicorn
python-dotenv
sentence-transformers
langchain
langchain-google-genai
langchain-community
firebase-admin
pyjwt
python-multipart
qdrant-client
google-genai
